requirements.txt
ansible>=2.14.0
ansible-core>=2.14.0
torch>=2.0.0
transformers>=4.30.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
click>=8.1.3
rich>=13.3.5
huggingface-hub>=0.16.0
tqdm>=4.65.0
numpy>=1.24.0
sentencepiece>=0.1.99
paramiko>=3.2.0  # For SSH connections to Windows hosts
tomli>=2.0.0  # For TOML configuration file parsing
